UPD reorganize shared components
- Always export default and keep one component per file rule.
