+++ frontend/local/src/containers/patients/index.js
@@ -7,9 +7,10 @@ import classnames from "classnames"
 import { read, DEFAULT_WAITLIST_ID } from "shared/modules/config"
 import { RESOURCE_PATIENT_IDENTIFICATION, READ, WRITE } from "../../modules/validations"
 import { search, cardToObject } from "../../modules/discovery"
-import { PatientImage, PatientCard } from "shared/containers/patient"
+import PatientImage from "shared/containers/patient/image"
+import PatientCard from "shared/containers/patient/card"
 import Spinner from "shared/containers/spinner"
-import { CodeTitle } from "shared/containers/codes"
+import CodeTitle from "shared/containers/codes/title"
 
 import { ReactComponent as SearchIcon } from "shared/icons/search.svg"
 import { ReactComponent as SearchActiveIcon } from "shared/icons/search-active.svg"
