+++ frontend/local/src/containers/waitlist/detail/add-diagnosis.js
@@ -8,7 +8,7 @@ import { searchCodes } from "shared/modules/codes"
 import { update as updateWaitlistItem, resetIndicators } from "../../../modules/waitlist"
 import { saveConsultation } from "../../../modules/patient"
 import Modal from "shared/containers/modal"
-import Patient from "shared/containers/patient"
+import Patient from "shared/containers/patient/card"
 import { renderInput, renderTextarea, renderReactSelect, validateRequired } from "shared/forms/renderField"
 import { cardToObject } from "../../../modules/discovery"
 import Spinner from "shared/containers/spinner"
